ft: Add serializer class
Add a serializer class to provide a way of serializing
and deserializing the snippet instances into `json`
representation as well as any others that may be needed.

Added the `create` and `update` methods to define how fully
fledged instances are created or modified when calling
`serializer.save()`

The `{'base_template': 'textarea.html'} flag is used to control
how the browsable API will be displayed.