ft: Add format suffix
Added format suffixes to API endpoints in views module to take
advantage of having responses that are not hardwired to a single
content type.

Added a set of `format_suffix_patterns` to urls module. This
gives us a simple clean way of referring to a singel format.

This gives us URLs that explicitly refer to a given format