Update serializer
Updated serializer to reflect association with the user
that created them.

The `source` argument controls which attribute is used
to populate a field, and can point an any attribute on
the serialized instance.

Using the dotted notation, it will traverse the given
attributes.

`ReadOnlyField` is always read-only and will be used for
serialized representations not updating model instances
when they are deserialized.

`CharField(read_only=True)` can also be used instead.