Modify serializers to extend `HyperlinkedModelSerializer`
Modified serializers to extend `HyperlinkedModelSerializer`
instead of the previous `ModelSerializer`.

`HyperlinkedModelSerializer` differs from `ModelSerializer`
in that it does not include the `id` field by default,
it includes the `url` field (using `HyperlinkedIdentityField`)
and relationships use the `HyperlinkedIdentityField` instead
of `PrimaryKeyRelatedField`.

Added `highlight` field which is the same type as the `url` field
except that it points to the `snippet-highlight` url pattern
instead of `snippet-detail` url pattern.

We indicate on the `highlight` field that any format suffixed
hyperlinks it returns should use `.html` suffix