Refactor views to use ViewSets
Refactored the user related views into one `UserViewSet`.
We used the `ReadOnlyModelViewSet` class to automatically
provide the default read-only operations.

Refactored the snippet related views into one `SnippetViewSet`
We used the `ModelViewSet` in order to get the complete set
of default read and write operations.

We used the `@action` decorator to create a custom action named
`highlight`. This decorator can be used to add any custom
endpoints that don't fit into the standard `create`/`update`/
`delete` style.

Custom actions which use the `@action` decorator will respond to
`GET` requests by default. The `methods` argument is used if we
want an action that responds to `POST` requests.

URLs for custom actions by default depend on the method name itself.
To change the way the url should be constructed, we can include `url_path`
as a decorator keyword argument.