Use `Router` class to wire views and urls conventions
Refactored urls module to use the `Router` class by
registering the appropriate view sets with the router

We provide two arguments - the URL prefix for the views,
and the viewset itself.

The `DefaultRouter` class automatically creates the API
root view for us, so we can now delete the `api_root` method
from the `views` module.